Player y NPC se mueven pero falta que el NPC tenga un % de moverse libremente o no

+++ FGD/include/Player.h
@@ -8,7 +8,7 @@ class Player : public Character {
 
  public:
     Player();
-    Player(BITMAP ***animations, int health, int damage, int speed, int shield, int x, int y, int height, int width);
+    Player(BITMAP ***animations, int health, int damage, double speed, int shield, int x, int y, int height, int width);
     //virtual ~Player();
     void keyboard();
