+++ FGD/src/Character.cpp
@@ -11,7 +11,7 @@ Character::Character()
     this->health = health;
 }*/
 
-Character::Character(BITMAP ***animations, int health, int damage, int speed, int shield, int x, int y, int height, int width):
+Character::Character(BITMAP ***animations, int health, int damage, double speed, int shield, int x, int y, int height, int width):
                     Drawable(animations, x, y, height, width)
 {
     this->health = health;
